Updated to use Model Downloader API and new firebase deps
