+++ DigitClassifier/GoogleService-Info.plist
@@ -2,22 +2,18 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-	<key>CLIENT_ID</key>
-	<string>1069647793992-08nvkm48d7r9ljov4ghheepknvr3a2nn.apps.googleusercontent.com</string>
-	<key>REVERSED_CLIENT_ID</key>
-	<string>com.googleusercontent.apps.1069647793992-08nvkm48d7r9ljov4ghheepknvr3a2nn</string>
 	<key>API_KEY</key>
-	<string>AIzaSyDvpfIZH2EoB-u6LFHJLUwXa-oTg9rn11Q</string>
+	<string>AIzaSyDV5hoxY2xelWnujvyz2AgAg_uBZ9j0ZdA</string>
 	<key>GCM_SENDER_ID</key>
-	<string>1069647793992</string>
+	<string>285317619820</string>
 	<key>PLIST_VERSION</key>
 	<string>1</string>
 	<key>BUNDLE_ID</key>
-	<string>com.firebase.uidemo</string>
+	<string>com.google.examples.DigitClassifier</string>
 	<key>PROJECT_ID</key>
-	<string>testapp-5d356</string>
+	<string>fir-ml-ios-codelab-5572c</string>
 	<key>STORAGE_BUCKET</key>
-	<string>testapp-5d356.appspot.com</string>
+	<string>fir-ml-ios-codelab-5572c.appspot.com</string>
 	<key>IS_ADS_ENABLED</key>
 	<false></false>
 	<key>IS_ANALYTICS_ENABLED</key>
@@ -29,8 +25,6 @@
 	<key>IS_SIGNIN_ENABLED</key>
 	<true></true>
 	<key>GOOGLE_APP_ID</key>
-	<string>1:1069647793992:ios:91eecf4730fc920b</string>
-	<key>DATABASE_URL</key>
-	<string>https://testapp-5d356.firebaseio.com</string>
+	<string>1:285317619820:ios:7ea75d9fcee435713ef849</string>
 </dict>
 </plist>
